FEATURE: extracts existing states logic into model-transformer class
